requirements.txt
langchain
pinecone-client
sentence-transformers
transformers
torch
PyMuPDF
python-dotenv
fpdf